pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.11.0",
    "chromadb>=1.3.4",
    "isort>=7.0.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.2",
    "pytest>=9.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "tavily-python>=0.7.12",
]
